requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
requests==2.31.0
python-dotenv==1.0.0
